implement warranty() logic, add functional & integration tests. add constraints endpoint

+++ pplansvc/pplans/flask/app.py
@@ -6,16 +6,21 @@ from morus.flask.app import (
     configured_app as morus_app,
     parse_args as morus_arg_parser
 )
+from morus.logging import getLogger
 
 from pplans.flask.blueprints import warranties_api
 from pplans.models import db
 from pplans.warranty import create_demo_data
 
+log = getLogger(__name__)
 
-log = logging.getLogger(__name__)
+# if README instructions were followed to create db & user, dsn will be:
+DEFAULT_DSN = "postgresql://testuser:testpass@localhost:5432/testdb"
 
 # example of extending argparser for specific service
 ConfiguredAppArgParser.add_argument("--dsn", required=True, help="DSN string")
+ConfiguredAppArgParser.add_argument("--testing", action="store_true", default=False,
+                                    help="create fresh copy of db with test data")
 
 def parse_args(parser=ConfiguredAppArgParser):
     args = morus_arg_parser(parser=parser)
@@ -23,8 +28,8 @@ def parse_args(parser=ConfiguredAppArgParser):
     return args
 
 # decorate morus_app to init db & register blueprint(s)
-def configured_app(import_name, dsn, debug=False, config_module=None,
-                   profile=False, proxy_fix=False):
+def configured_app(import_name, dsn, debug=False, testing=False,
+                   config_module=None, profile=False, proxy_fix=False):
     app = morus_app(import_name, debug=debug, config_module=config_module,
                     profile=profile, proxy_fix=proxy_fix)
     app.config["SQLALCHEMY_DATABASE_URI"] = dsn
@@ -34,8 +39,9 @@ def configured_app(import_name, dsn, debug=False, config_module=None,
     db.init_app(app)
     log.debug("configured_app: {}".format(app))
     # for demo purposes..
-    db.drop_all()
-    db.create_all()
-    create_demo_data()
+    if testing:
+        db.drop_all()
+        db.create_all()
+        create_demo_data()
     return app
